\Total:                   40.0 of 70.0\

+++ lex.yy.cc
@@ -1038,7 +1038,7 @@ YY_RULE_SETUP
   while(!results.empty()) {
     char temp = *results.back();
     // fprintf(stderr, "%c", temp);
-    myunputc(results.back());
+    myunputc((int)results.back());
     results.pop_back();
   }
